+++ lib/ndr_dev_support/capistrano/assets.rb
@@ -4,8 +4,8 @@ Capistrano::Configuration.instance(:must_exist).load do
     task :configure_assets do
       asset_script = fetch(:asset_script, <<~SHELL)
         set -e
-        ruby -ryaml -e "puts YAML.dump('production' => { 'secret_key_base' => 'compile_me' })" > config/secrets.yml
-        ruby -ryaml -e "puts YAML.dump('production' => { 'adapter' => 'placeholder' })" > config/database.yml
+        ruby -e "require 'yaml'; puts YAML.dump('production' => { 'secret_key_base' => 'compile_me' })" > config/secrets.yml
+        ruby -e "require 'yaml'; puts YAML.dump('production' => { 'adapter' => 'placeholder' })" > config/database.yml
         RAILS_ENV=production bundle exec rake assets:clobber assets:precompile
         rm config/secrets.yml config/database.yml
       SHELL